requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langgraph==0.2.60
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-google-genai==2.0.7
langchain-chroma==0.1.4
chromadb==0.5.23
python-dotenv==1.0.1
sse-starlette==2.2.1
pydantic==2.10.4
pydantic-settings==2.7.0
google-generativeai==0.8.3
